app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
rq
pydantic
pytest
httpx
pydantic-settings
alembic
tree-sitter
tree-sitter-java
tree-sitter-cpp
websockets
pyjwt
